+++ src/main/java/frc/robot/BuildConstants.java
@@ -5,12 +5,12 @@ public final class BuildConstants {
   public static final String MAVEN_GROUP = "";
   public static final String MAVEN_NAME = "Offseason2026-1";
   public static final String VERSION = "unspecified";
-  public static final int GIT_REVISION = 8;
-  public static final String GIT_SHA = "ef5a8ea278574ddfe27ab8405852b293429db437";
-  public static final String GIT_DATE = "2026-05-05 20:19:55 EDT";
+  public static final int GIT_REVISION = 9;
+  public static final String GIT_SHA = "d88d444d1825097b316414f9cc565bf90ddbe0e2";
+  public static final String GIT_DATE = "2026-05-19 16:33:38 EDT";
   public static final String GIT_BRANCH = "intakeDeploySubsystem";
-  public static final String BUILD_DATE = "2026-05-19 16:06:58 EDT";
-  public static final long BUILD_UNIX_TIME = 1779221218715L;
+  public static final String BUILD_DATE = "2026-05-19 17:03:39 EDT";
+  public static final long BUILD_UNIX_TIME = 1779224619701L;
   public static final int DIRTY = 1;
 
   private BuildConstants() {}
